+++ packages/@expo/fingerprint/src/Fingerprint.ts
@@ -1,6 +1,6 @@
 import { dedupSources } from './Dedup';
 import type { Fingerprint, FingerprintSource, Options } from './Fingerprint.types';
-import { normalizeOptions } from './Options';
+import { normalizeOptionsAsync } from './Options';
 import { sortSources } from './Sort';
 import { createFingerprintFromSourcesAsync } from './hash/Hash';
 import { getHashSourcesAsync } from './sourcer/Sourcer';
@@ -12,7 +12,7 @@ export async function createFingerprintAsync(
   projectRoot: string,
   options?: Options
 ): Promise<Fingerprint> {
-  const opts = normalizeOptions(options);
+  const opts = await normalizeOptionsAsync(projectRoot, options);
   const sources = await getHashSourcesAsync(projectRoot, opts);
   const normalizedSources = sortSources(dedupSources(sources, projectRoot));
   const fingerprint = await createFingerprintFromSourcesAsync(normalizedSources, projectRoot, opts);